+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/content/kinetics/encased_blocks/cogwheel/BnbEncasedFlangedCogBlock.java
@@ -1,7 +1,6 @@
-package com.kipti.bnb.content.kinetics.encased_blocks.cogwheel_chain;
+package com.kipti.bnb.content.kinetics.encased_blocks.cogwheel;
 
 import com.kipti.bnb.content.kinetics.cogwheel_chain.block.IFlangedCogWheel;
-import com.kipti.bnb.content.kinetics.encased_blocks.BnbEncasedCogwheelBlock;
 import com.kipti.bnb.registry.content.BnbBlockEntities;
 import com.simibubi.create.content.kinetics.base.KineticBlockEntity;
 import com.simibubi.create.content.kinetics.simpleRelays.SimpleKineticBlockEntity;
@@ -11,7 +10,6 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.LevelEvent;
@@ -20,13 +18,14 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
+import org.jspecify.annotations.NonNull;
 
 import java.util.function.Supplier;
 
-public class BnbEncasedEmptyFlangedGearBlock extends BnbEncasedCogwheelBlock implements IFlangedCogWheel {
+public class BnbEncasedFlangedCogBlock extends BnbEncasedCogwheelBlock implements IFlangedCogWheel {
     private final Supplier<Block> flangedCogwheel;
 
-    public BnbEncasedEmptyFlangedGearBlock(final Properties properties, final boolean large, final Supplier<Block> casing, final Supplier<Block> flangedCogwheel) {
+    public BnbEncasedFlangedCogBlock(final Properties properties, final boolean large, final Supplier<Block> casing, final Supplier<Block> flangedCogwheel) {
         super(properties, large, casing);
         this.flangedCogwheel = flangedCogwheel;
     }
@@ -38,28 +37,28 @@ public class BnbEncasedEmptyFlangedGearBlock extends BnbEncasedCogwheelBlock imp
         context.getLevel()
                 .levelEvent(LevelEvent.PARTICLES_DESTROY_BLOCK, context.getClickedPos(), Block.getId(state));
         KineticBlockEntity.switchToBlockState(context.getLevel(), context.getClickedPos(),
-                flangedCogwheel.get().defaultBlockState()
+                this.flangedCogwheel.get().defaultBlockState()
                         .setValue(AXIS, state.getValue(AXIS)));
         return InteractionResult.SUCCESS;
     }
 
     @Override
-    public ItemStack getCloneItemStack(final BlockState state, final HitResult target, final LevelReader level, final BlockPos pos, final Player player) {
+    public @NonNull ItemStack getCloneItemStack(final BlockState state, final HitResult target, final LevelReader level, final BlockPos pos, final Player player) {
         if (target instanceof final BlockHitResult blockHitResult)
             return blockHitResult.getDirection()
-                            .getAxis() != getRotationAxis(state)
-                    ? flangedCogwheel.get().asItem().getDefaultInstance()
-                    : getCasing().asItem().getDefaultInstance();
+                    .getAxis() != this.getRotationAxis(state)
+                    ? this.flangedCogwheel.get().asItem().getDefaultInstance()
+                    : this.getCasing().asItem().getDefaultInstance();
         return super.getCloneItemStack(state, target, level, pos, player);
     }
 
     @Override
     public ItemRequirement getRequiredItems(final BlockState state, final BlockEntity be) {
-        return ItemRequirement.of(flangedCogwheel.get().defaultBlockState(), be);
+        return ItemRequirement.of(this.flangedCogwheel.get().defaultBlockState(), be);
     }
 
     @Override
     public BlockEntityType<? extends SimpleKineticBlockEntity> getBlockEntityType() {
-        return BnbBlockEntities.SIMPLE_KINETIC.get();
+        return this.isLarge ? BnbBlockEntities.ENCASED_LARGE_FLANGED_COGWHEEL.get() : BnbBlockEntities.ENCASED_FLANGED_COGWHEEL.get();
     }
 }