+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/mixin/compat/cogwheel_material/createadditionallogistics/PackageAcceleratorVisualMixin.java
@@ -36,7 +36,7 @@ public abstract class PackageAcceleratorVisualMixin extends AbstractBlockEntityV
     public Instancer<RotatingInstance> bnb$materialCog(final InstancerProvider provider, final InstanceType<RotatingInstance> type, final Model model, final Operation<Instancer<RotatingInstance>> original) {
         if (this.bnb$materialState == null)
             this.bnb$materialState = new State();
-        return this.bnb$materialInstancer(provider, type, model, original, null, null, this.bnb$materialState, this.blockEntity, this.blockState);
+        return this.bnb$materialInstancer(provider, type, model, null, null, this.bnb$materialState, this.blockEntity, this.blockState);
     }
 
     @Inject(method = "update", at = @At("HEAD"))