+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/mixin/compat/cogwheel_material/createcasing/CreativeCogwheelVisualMixin.java
@@ -41,7 +41,6 @@ public abstract class CreativeCogwheelVisualMixin extends AbstractBlockEntityVis
         return this.bnb$materialInstancer(provider,
                 type,
                 model,
-                original,
                 Direction.UP,
                 Direction.fromAxisAndDirection(this.blockState.getValue(BlockStateProperties.AXIS), Direction.AxisDirection.POSITIVE),
                 this.bnb$materialState,