+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/registry/content/BnbBlockEntities.java
@@ -7,6 +7,8 @@ import com.kipti.bnb.content.kinetics.chain_pulley.ChainPulleyBlockEntity;
 import com.kipti.bnb.content.kinetics.chain_pulley.ChainPulleyRenderer;
 import com.kipti.bnb.content.kinetics.cogwheel_carriage.block.CogwheelChainCarriageBlockEntity;
 import com.kipti.bnb.content.kinetics.cogwheel_chain.migration.MigratingSimpleKineticBlockEntity;
+import com.kipti.bnb.content.kinetics.encased_blocks.cogwheel.BnbEncasedFlangedCogRenderer;
+import com.kipti.bnb.content.kinetics.encased_blocks.cogwheel.BnbEncasedFlangedCogVisual;
 import com.kipti.bnb.content.kinetics.flywheel_bearing.FlywheelBearingBlockEntity;
 import com.kipti.bnb.content.kinetics.flywheel_bearing.FlywheelBearingBlockEntityRenderer;
 import com.kipti.bnb.content.kinetics.gigantic_cogwheel.GiganticCogwheelBlockEntity;
@@ -67,19 +69,41 @@ public class BnbBlockEntities {
                     BnbExtraEncasedBlocks.INDUSTRIAL_IRON_ENCASED_SHAFT,
                     BnbExtraEncasedBlocks.WEATHERED_IRON_ENCASED_SHAFT,
                     BnbDecorativeBlocks.METAL_TRUSS_SHAFT
-//                    BnbSpecialEncasedBlocks.INDUSTRIAL_GRATING_PANEL
-//                    BnbDecorativeBlocks.INDUSTRIAL_TRUSS_ENCASED_SHAFT
             )
             .renderer(() -> ShaftRenderer::new)
             .register();
-//
-//    public static final BlockEntityEntry<FluidPipeBlockEntity> ENCASED_PIPE = REGISTRATE
-//            .blockEntity("encased_pipe", FluidPipeBlockEntity::new)
-//            .validBlocks(
-//                    BnbSpecialEncasedBlocks.INDUSTRIAL_GRATING_PANEL_PIPE,
-//                    BnbDecorativeBlocks.INDUSTRIAL_TRUSS_ENCASED_PIPE
-//            )
-//            .register();
+
+    public static final BlockEntityEntry<SimpleKineticBlockEntity> ENCASED_FLANGED_COGWHEEL = REGISTRATE
+            .blockEntity("encased_flanged_cogwheel", SimpleKineticBlockEntity::new)
+            .visual(() -> BnbEncasedFlangedCogVisual::small, false)
+            .validBlocks(BnbEncasedBlockLists.ENCASED_FLANGED_COGWHEEL.toArray())
+            .renderer(() -> BnbEncasedFlangedCogRenderer::small)
+            .register();
+
+    public static final BlockEntityEntry<SimpleKineticBlockEntity> ENCASED_LARGE_FLANGED_COGWHEEL = REGISTRATE
+            .blockEntity("encased_large_flanged_cogwheel", SimpleKineticBlockEntity::new)
+            .visual(() -> BnbEncasedFlangedCogVisual::large, false)
+            .validBlocks(BnbEncasedBlockLists.ENCASED_LARGE_FLANGED_COGWHEEL.toArray())
+            .renderer(() -> BnbEncasedFlangedCogRenderer::large)
+            .register();
+
+    public static final BlockEntityEntry<SimpleKineticBlockEntity> SIMPLE_KINETIC = REGISTRATE.blockEntity(
+                    "simple_kinetic",
+                    SimpleKineticBlockEntity::new
+            )
+            .visual(
+                    () -> (context, blockEntity, partialTick) ->
+                            new SingleAxisRotatingVisual<>(
+                                    context, blockEntity, partialTick,
+                                    Models.partial(GenericBlockEntityRenderModels.REGISTRY.get(blockEntity.getBlockState().getBlock()))
+                            ), true
+            )
+            .validBlocks(
+                    BnbKineticBlocks.SMALL_FLANGED_COGWHEEL,
+                    BnbKineticBlocks.LARGE_FLANGED_COGWHEEL
+            )
+            .renderer(() -> KineticBlockEntityRenderer::new)
+            .register();
 
     public static final BlockEntityEntry<SimpleKineticBlockEntity> ENCASED_COGWHEEL = REGISTRATE
             .blockEntity("encased_cogwheel", SimpleKineticBlockEntity::new)
@@ -137,26 +161,6 @@ public class BnbBlockEntities {
             )
             .validBlocks(BnbKineticBlocks.SMALL_FLANGED_COGWHEEL, BnbKineticBlocks.LARGE_FLANGED_COGWHEEL)
             .validBlocks(AllBlocks.COGWHEEL, AllBlocks.LARGE_COGWHEEL)
-            .renderer(() -> KineticBlockEntityRenderer::new)
-            .register();
-
-    public static final BlockEntityEntry<SimpleKineticBlockEntity> SIMPLE_KINETIC = REGISTRATE.blockEntity(
-                    "simple_kinetic",
-                    SimpleKineticBlockEntity::new
-            )
-            .visual(
-                    () -> (context, blockEntity, partialTick) ->
-                            new SingleAxisRotatingVisual<>(
-                                    context, blockEntity, partialTick,
-                                    Models.partial(GenericBlockEntityRenderModels.REGISTRY.get(blockEntity.getBlockState().getBlock()))
-                            ), true
-            )
-            .validBlocks(
-                    BnbKineticBlocks.SMALL_FLANGED_COGWHEEL,
-                    BnbKineticBlocks.LARGE_FLANGED_COGWHEEL
-            )
-            .validBlocks(BnbEncasedBlockLists.ENCASED_LARGE_FLANGED_COGWHEEL.toArray())
-            .validBlocks(BnbEncasedBlockLists.ENCASED_FLANGED_COGWHEEL.toArray())
             .renderer(() -> KineticBlockEntityRenderer::new)
             .register();